Move Monte Carlo plotting class to plotting_MC.py / Update Bar Chart to Plotly
