+++ app/utils/plotting.py
@@ -22,13 +22,6 @@ import matplotlib.pyplot as plt
 import matplotlib.ticker as ticker
 import numpy as np
 import pandas as pd
-from matplotlib.patches import Circle, RegularPolygon
-from matplotlib.path import Path
-from matplotlib.projections import register_projection
-from matplotlib.projections.polar import PolarAxes
-from matplotlib.spines import Spine
-from matplotlib.transforms import Affine2D
-import plotly.express as px
 import plotly.graph_objects as go
 from plotly.subplots import make_subplots
 import plotly.subplots as sp
@@ -145,80 +138,7 @@ def _decile(
     return ax
 
 
-def radar_factory(num_vars, frame="circle"):
-    """
-    Create a radar chart with `num_vars` axes.
 
-    This function creates a RadarAxes projection and registers it.
-
-    Parameters
-    ----------
-    num_vars : int
-        Number of variables for radar chart.
-    frame : {'circle', 'polygon'}
-        Shape of frame surrounding axes.
-
-    """
-    # calculate evenly-spaced axis angles
-    theta = np.linspace(0, 2 * np.pi, num_vars, endpoint=False)
-
-    class RadarAxes(PolarAxes):
-        name = "radar"
-        # use 1 line segment to connect specified points
-        RESOLUTION = 1
-
-        def __init__(self, *args, **kwargs):
-            super().__init__(*args, **kwargs)
-            # rotate plot such that the first axis is at the top
-            self.set_theta_zero_location("N")
-
-        def fill(self, *args, closed=True, **kwargs):
-            """Override fill so that line is closed by default"""
-            return super().fill(closed=closed, *args, **kwargs)
-
-        def plot(self, *args, **kwargs):
-            """Override plot so that line is closed by default"""
-            lines = super().plot(*args, **kwargs)
-            for line in lines:
-                self._close_line(line)
-
-        def _close_line(self, line):
-            x, y = line.get_data()
-            # FIXME: markers at x[0], y[0] get doubled-up
-            if x[0] != x[-1]:
-                x = np.append(x, x[0])
-                y = np.append(y, y[0])
-                line.set_data(x, y)
-
-        def set_varlabels(self, labels):
-            self.set_thetagrids(np.degrees(theta), labels)
-
-        def _gen_axes_patch(self):
-            # The Axes patch must be centered at (0.5, 0.5) and of radius 0.5
-            # in axes coordinates.
-            if frame == "circle":
-                return Circle((0.5, 0.5), 0.5)
-            elif frame == "polygon":
-                return RegularPolygon((0.5, 0.5), num_vars, radius=0.5, edgecolor="k")
-            else:
-                raise ValueError("Unknown value for 'frame': %s" % frame)
-
-        def _gen_axes_spines(self):
-            if frame == "circle":
-                return super()._gen_axes_spines()
-            elif frame == "polygon":
-                # spine_type must be 'left'/'right'/'top'/'bottom'/'circle'.
-                spine = Spine(axes=self, spine_type="circle", path=Path.unit_regular_polygon(num_vars))
-                # unit_regular_polygon gives a polygon of radius 1 centered at
-                # (0, 0) but we want a polygon of radius 0.5 centered at (0.5,
-                # 0.5) in axes coordinates.
-                spine.set_transform(Affine2D().scale(0.5).translate(0.5, 0.5) + self.transAxes)
-                return {"polar": spine}
-            else:
-                raise ValueError("Unknown value for 'frame': %s" % frame)
-
-    register_projection(RadarAxes)
-    return theta
 
 
 class PLOT:
@@ -1203,146 +1123,3 @@ class PLOT:
 
         return fig
 
-class PLOT_MC:
-    def __init__(self, dic):
-        self.fig1 = self.radar_montecarlo(dic)
-        self.fig2 = self.bar_with_uncertainty(dic)
-
-        self.figs = [
-            {
-                "title": "Uncertainty Analysis - Monte Carlo (Radar Chart)",
-                "plot": self.fig1,
-                "type" : "matplotlib"
-            }, 
-            {
-                "title": "Uncertainty Analysis - Monte Carlo (Bar Chart)",
-                "plot": self.fig2,
-                "type" : "matplotlib"
-            }
-        ]
-
-    def radar_montecarlo(self, dic):
-        N = len(dic["EI_name"])
-        theta = radar_factory(N, frame="polygon")
-
-        excel_LCA = pd.ExcelFile(os.path.join(dic["LCA_path"], dic["filename_result_EI_MC"]))
-        df_LCA = pd.read_excel(excel_LCA, 0)
-        excel_LCA.close()
-
-        mean = df_LCA.loc[:, "Mean"]
-        sd = df_LCA.loc[:, "SD"]
-
-        mean = mean.to_numpy()
-        sd = sd.to_numpy()
-
-        mean_sum_norm = 100 * mean / mean
-
-        min_sum_norm = 100 * (mean - 2 * sd) / mean
-        max_sum_norm = 100 * (mean + 2 * sd) / mean
-
-        categories = [f"{name} ({unit})" for name, unit in zip(dic["EI_name"], dic["LCIA_unit"])]
-
-        # Define the data for each category
-        max_values = [max_sum_norm[i] for i in range(len(categories))]
-        mean_values = [mean_sum_norm[i] for i in range(len(categories))]
-        min_values = [min_sum_norm[i] for i in range(len(categories))]
-
-        # Package the data into a structured format
-        data = [
-            categories,  # List of categories
-            (
-                "Uncertainty Analysis - Monte Carlo",  # Title for the analysis
-                [
-                    max_values,  # Max values for each category
-                    mean_values,  # Mean values for each category
-                    min_values,  # Min values for each category
-                ],
-            ),
-        ]
-
-        spoke_labels = data.pop(0)
-
-        colors = ["m", "w", "w"]
-        colors_line = ["mo", "m", "m."]
-
-        title, case_data = data[0]
-
-        fig, ax = plt.subplots(subplot_kw=dict(projection="radar"), figsize=(7, 5))
-        fig.subplots_adjust(top=0.85, bottom=0.05)
-
-        ax.set_title(
-            title, fontsize=20, weight="bold", horizontalalignment="center", verticalalignment="center", pad=20
-        )
-        alphas = [0.35, 0.25, 1]
-        for d, color, alp, c_l in zip(case_data, colors, alphas, colors_line):
-            line = ax.plot(theta, d, c_l)
-            ax.fill(theta, d, facecolor=color, alpha=alp, label="_nolegend_")
-
-        ax.set_varlabels(spoke_labels)
-
-        # add legend relative to top-left plot
-        labels = ("μ+2σ", "Mean (μ)", "μ-2σ")
-        legend = ax.legend(labels, loc="upper center", bbox_to_anchor=(0.5, -0.1), ncol=3, fontsize=14)
-
-        return fig
-
-    def bar_with_uncertainty(self, dic):
-        N = len(dic["EI_name"])
-
-        # Charger les données
-        excel_LCA = pd.ExcelFile(os.path.join(dic["LCA_path"], dic["filename_result_EI_MC"]))
-        df_LCA = pd.read_excel(excel_LCA, 0)
-
-        excel_LCA.close()
-
-        # Extraire les moyennes et les écarts-types
-        mean = df_LCA.loc[:, "Mean"].to_numpy()
-        sd = df_LCA.loc[:, "SD"].to_numpy()
-
-        # Normalisation
-        mean_sum_norm = 100 * mean / mean
-        min_sum_norm = 100 * (mean - 2 * sd) / mean
-        max_sum_norm = 100 * (mean + 2 * sd) / mean
-
-        # Labels des catégories
-        categories = df_LCA.loc[:, "Method"].to_numpy()
-
-        combined_labels = [f"{name} ({unit})" for name, unit in zip(categories, dic["LCIA_unit"])]
-
-        # Création du graphique à barres
-        fig, ax = plt.subplots(figsize=(10, 6))
-
-        # Barres de moyenne
-        yerr = [min_sum_norm, max_sum_norm]
-
-        # Calcul des barres d'erreur avec correction des valeurs négatives
-        lower_error = np.clip(mean_sum_norm - min_sum_norm, 0, None)
-        upper_error = np.clip(max_sum_norm - mean_sum_norm, 0, None)
-
-        bars = ax.bar(
-            combined_labels,
-            mean_sum_norm,
-            yerr=[lower_error, upper_error],
-            capsize=5,
-            color="plum",
-            edgecolor="none",
-            ecolor="midnightblue",
-        )
-
-        # Ajouter les labels
-        ax.set_ylabel("Normalized Value (%)", fontsize=14)
-        ax.set_title("Uncertainty Analysis - Monte Carlo", weight="bold", fontsize=20)
-        ax.tick_params(axis="x", rotation=45, labelsize=14)
-        ax.tick_params(axis="y", labelsize=14)
-        ax.set_xticks(np.arange(len(categories)))  # Assurez-vous que les ticks sont au bon endroit
-        ax.set_xticklabels(combined_labels, rotation=45, ha="right", fontsize=12)
-
-        # Add grid
-        ax.yaxis.grid(True, linestyle="--", linewidth=0.7, color="grey")
-
-        # Ajouter la légende
-        ax.legend(["Uncertainty (μ±2σ)"], loc="upper center", bbox_to_anchor=(0.5, -0.4), fontsize=14, ncol=1)
-
-        plt.tight_layout()
-
-        return fig
